# feat: Single-compilation CI optimization for 4x faster builds