--- a/scripts/test-single-compilation.sh
+++ b/scripts/test-single-compilation.sh
@@ -1,0 +1,99 @@
+#!/bin/bash
+set -e
+
+# Test script to validate single-compilation strategy locally
+# This simulates the optimized CI pipeline approach
+
+echo "🧪 Testing Single-Compilation Strategy Locally"
+echo "============================================="
+
+# Colors for output
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+YELLOW='\033[1;33m'
+BLUE='\033[0;34m'
+NC='\033[0m' # No Color
+
+# Track timing
+start_time=$(date +%s)
+
+echo -e "${BLUE}Phase 1: Single Comprehensive Build${NC}"
+echo "Building all targets once..."
+
+build_start=$(date +%s)
+
+# Install nextest if not available
+if ! command -v cargo-nextest &> /dev/null; then
+    echo "Installing cargo-nextest..."
+    cargo install cargo-nextest --locked
+fi
+
+# Single comprehensive build (matches optimized CI)
+echo "🏗️ Building dev profile with all features..."
+cargo build --all-features
+
+echo "🏗️ Building test binaries..."  
+cargo build --all-features --tests
+
+echo "🏗️ Building release binary..."
+cargo build --release --bin kotadb
+
+build_end=$(date +%s)
+build_duration=$((build_end - build_start))
+
+echo -e "${GREEN}✅ Build completed in ${build_duration} seconds${NC}"
+
+echo -e "${BLUE}Phase 2: Artifact-based Testing${NC}"
+echo "Running all tests from single compilation..."
+
+test_start=$(date +%s)
+
+# Unit tests (using pre-compiled artifacts)
+echo "🚀 Running unit tests (no recompilation)..."
+cargo nextest run --lib --all-features --no-fail-fast
+cargo test --doc --all-features
+
+# Integration tests (simulating matrix execution)
+echo "🧪 Running integration tests (simulating parallel execution)..."
+echo "  - Core storage tests..."
+cargo nextest run --test file_storage_integration_test
+cargo nextest run --test data_integrity_test
+
+echo "  - API system tests..."
+cargo nextest run --test http_server_integration_test
+cargo nextest run --test cli_defaults_validation_test
+
+# E2E tests
+echo "🎯 Running E2E tests..."
+cargo nextest run --test e2e_integration_test
+
+test_end=$(date +%s)
+test_duration=$((test_end - test_start))
+
+end_time=$(date +%s)
+total_duration=$((end_time - start_time))
+
+echo -e "${GREEN}✅ All tests completed in ${test_duration} seconds${NC}"
+
+echo ""
+echo "📊 Performance Analysis:"
+echo "========================"
+echo -e "Build time (single compilation): ${YELLOW}${build_duration}s${NC}"
+echo -e "Test execution time: ${YELLOW}${test_duration}s${NC}"
+echo -e "Total pipeline time: ${YELLOW}${total_duration}s${NC}"
+
+echo ""
+echo "🎯 Single-Compilation Benefits:"
+echo "- ✅ One comprehensive build covers all test needs"
+echo "- ✅ Zero recompilation overhead between test phases"
+echo "- ✅ Tests run from pre-compiled binaries (faster)"
+echo "- ✅ Same artifacts used across all test categories"
+
+echo ""
+echo -e "${GREEN}🚀 Single-compilation strategy validation complete!${NC}"
+
+# Compare with traditional approach (informational)
+echo ""
+echo "💡 Traditional CI would compile separately for each matrix job."
+echo "   Estimated traditional overhead: ~4x build time = $((build_duration * 4))s"
+echo -e "   Single-compilation savings: ${GREEN}~$((build_duration * 3))s per CI run${NC}"